Replace stdin command interface with a named pipe one
The command interface currently supports reading from
stdin or a network socket.
Using stdin is inconvenient and the socket is potentially
dangerous since you have to protect access to the port
from the outside.

Replace the stdin interface with a named pipe. The filename
of the pipe can the configured through the config and
securing the pipe is just a matter of settings correct
permissions.