gfx/exynos: add software framebuffer support
This swfb API is different from the one in upstream RetroArch.
I decided that I need at least the possibility to change
pixel formats on-the-fly and to specify a blitting rectangle, in
case the core only wants to commit a subregion of the
framebuffer (like e.g. overscan removal in Snes9x).

Missing features (coming in a later commit):
- menu rendering
- font rendering
- soft filter