autosave: refactor and improve exit behaviour of autosave thread
Currently when the autosave thread is waiting for the condition
variable to time out, and core code calls autosave_free(), the
while loop is immediately left.

Change the logic so that if the condition variable is signaled, the
loop is run for another iteration. The rationale is to do a last
check/save before stopping the autosave thread.